moved presence away from discord
